+++ example_api_calls/delete-policy.sh
@@ -3,4 +3,4 @@
 POLICY_NAME="default"
 API_KEY=""
 
-curl -s -X DELETE "http://localhost:8080/api/policies/${POLICY_NAME}" -H "Authorization: Bearer ${API_KEY}"
+curl -k -s -X DELETE "https://localhost:8080/api/policies/${POLICY_NAME}" -H "Authorization: Bearer ${API_KEY}"