+++ example_api_calls/filter.sh
@@ -3,7 +3,7 @@
 API_KEY=""
 
 # Create an SSN-only policy.
-curl -v -s -X POST "http://localhost:8080/api/policies" -H "Content-Type: application/json" -H "Authorization: Bearer ${API_KEY}" -d @./ssn.json
+curl -k -v -s -X POST "https://localhost:8080/api/policies" -H "Content-Type: application/json" -H "Authorization: Bearer ${API_KEY}" -d @./ssn.json
 
 # Apply the policy to text.
-curl -s -X POST "http://localhost:8080/api/filter?p=ssn" -H "Content-Type: text/plain" -H "Authorization: Bearer ${API_KEY}" -H "Accept: text/plain" -d'His SSN was 123-45-6789.'
+curl -k -s -X POST "https://localhost:8080/api/filter?p=ssn" -H "Content-Type: text/plain" -H "Authorization: Bearer ${API_KEY}" -H "Accept: text/plain" -d'His SSN was 123-45-6789.'