+++ example_api_calls/get-policies.sh
@@ -2,4 +2,4 @@
 
 API_KEY=""
 
-curl -s -X GET "http://localhost:8080/api/policies" -H "Accept: application/json" -H "Authorization: Bearer ${API_KEY}" | jq
+curl -k -s -X GET "https://localhost:8080/api/policies" -H "Accept: application/json" -H "Authorization: Bearer ${API_KEY}" | jq