+++ example_api_calls/get-policy.sh
@@ -3,4 +3,4 @@
 POLICY_NAME="default"
 API_KEY=""
 
-curl -s -X GET "http://localhost:8080/api/policies/${POLICY_NAME}" -H "Authorization: Bearer ${API_KEY}" -H "Accept: application/json" | jq
+curl -k -s -X GET "https://localhost:8080/api/policies/${POLICY_NAME}" -H "Authorization: Bearer ${API_KEY}" -H "Accept: application/json" | jq