+++ example_api_calls/save-policy.sh
@@ -2,4 +2,4 @@
 
 API_KEY=""
 
-curl -v -s -X POST "http://localhost:8080/api/policies" -H "Authorization: Bearer ${API_KEY}" -H "Content-Type: application/json"  -d @../distribution/policies/default.json
+curl -k -v -s -X POST "https://localhost:8080/api/policies" -H "Authorization: Bearer ${API_KEY}" -H "Content-Type: application/json"  -d @../distribution/policies/default.json